backend/pyproject.toml
[project]
name = "pyqo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-cors-headers>=4.6.0",
    "django-filter>=24.3",
    "django-geoposition-2>=0.4.0",
    "django>=5.1.3",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
    "drf-standardized-errors>=0.14.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "uuid>=1.30",
    "uvicorn>=0.32.0",
    "whitenoise>=6.8.2",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "pre-commit>=4.0.1",
    "python-dotenv>=1.0.1",
    "ruff>=0.7.3",
]
